requirements.txt
Pillow
pyrogram
tgcrypto
pyppeteer